revert(restore): withdraw --overwrite; it could not be made safe
--overwrite was implemented on this branch, put through four adversarial
passes, and is being withdrawn on the evidence of them. The flag remains and
reports; it no longer deletes anything.

Replacing a snapshot means deleting it before receiving its replacement. There
is no way around that: a received subvolume is read-only and cannot be renamed
OR moved -- verified on real btrfs, where even a move to a different parent
under the same name fails with EROFS -- and btrfs refuses to clear the
read-only flag while received_uuid is set, which is what incremental send
depends on. So there is no staging area and no atomic swap, and the destination
necessarily holds neither copy for the duration of the transfer.

The argument for accepting that window was that a restore only ever reads the
backup, so an interruption costs a retry rather than data. Each review round
found another way for it to cost data instead.

The window was wider than the transfer: the deletion happened before the lock
was taken, so a network failure that never moved a byte destroyed the copy. The
space check ran after the delete, so the copy was destroyed for a shortage the
tool already had the means to detect. Replacing one snapshot deleted its whole
ancestor chain, subvolumes the operator never named, while the consent they had
given spoke of one. A delete that silently declined -- a locked snapshot, a
refused sudo -- reported success, and the whole snapshot then streamed and
failed at the end with "File exists".

Each of those was fixed. The last one was not fixable in the same way: every
check of whether the backup could be delivered at all -- a corrupt stream, a
missing decompressor, an unsupported cipher -- lived inside the send, i.e.
downstream of the deletion. A corrupt backup therefore cost the operator the
last good copy, and the advice printed afterwards told them to run the same
command again, which failed identically every time. Reproduced on real btrfs
through the real CLI, with a control run proving the copy survives when the
check is hoisted.

That check IS hoisted, and stays: preflight_send exposes the read-side checks
send already makes, so every restore now finds an undeliverable backup before
streaming rather than during. What does not stay is the deletion it was
protecting.

The pattern is the reason for withdrawing rather than continuing. Four rounds,
each finding its worst defects inside the previous round's fixes, on a feature
whose failure mode is permanent loss of the copy the operator asked to refresh.
A safety property that needs four passes and is still yielding
deletion-before-verification is not a safety property.

The flag is kept as a reporting one rather than removed or made an error.
Accepting it and doing nothing is the defect this project has spent several
releases removing -- an option recognised and silently ignored -- and an
operator passing it to refresh a stale copy would believe it had been
refreshed. Making it an error would break working invocations for no gain,
since a restore that brings back the snapshots that ARE missing is still doing
its job. So it says plainly that it was not applied, names what it left alone,
and the restore continues.

Kept from the work: reading the destination under the prefix the source ended
up with, one _choose_parent shared by the dry-run preview and the real run,
relatedness that raw snapshots can answer, unorderable pairs degrading to a
full send, and the collision check that no longer reports "no collision" when
it could not read the destination at all.

A redesign belongs in 0.9.7, and the constraint to design against is the one
established here: the deletion must not precede the proof that the replacement
can be delivered.
